Crunch port: create engine input-table DDL; full schema now applies + plans clean
Ported 55 input/staging tables + views referenced by the crunch engine but
created by no migration (the known "input tables uncreated" data-layer gap):

  002b_dataload_tables.sql        15 dataload.* tables
  002c_xwalk_tables.sql           16 xwalk.* tables + xwalk.gradelevel
                                  + wageareatofips view
  002d_staging_tables.sql         "DMDC".* / "load_GFEBS".* / load_inventory.*
                                  / load_training.* (9)
  002e_payschedule_raw_tables.sql 11 "PaySchedule".* raw tables
  008d_input_views.sql            data.categorygroup/categorysubgroup/
                                  costswithdescriptions + xwalk.payplantype/
                                  ppxwalkgradelevel

Wired into init.sh (002b-e after 002; 008d after 008c). Full 000->008d now
applies from scratch: 275 tables / 164 views / 49 procs; zero referenced-but-
missing relations; all 48 proc calls + crunch.crunchall plan clean.

Real bugs surfaced once the tables existed, fixed here:
- xwalk.fips_wagearea: stale sqlproj DDL (single AmcosVersionId) vs consumers
  reading start/end (006e + source SP) and single (006j wageareatofips) -
  carry both as a superset (start/end nullable); flagged for real-data recon.
- xwalk.seriestosoc: stale DDL column Series -> occupationalseriesnumber
  (name every consumer + the source SP actually reads).
- xwalk.ppxwalkgradelevel view: 4 gradelevel BETWEEN band sites cast ::smallint
  (data.costs.gradelevel smallint vs varchar bands; mirrors SQL Server implicit
  varchar->int); payband unified to varchar to match warehouse.ppxwalk target.
- 006j populateppxwalk: gs_ses_gradelevel join cast ::smallint.

Structural port only; numeric validation still gated on real ETL data.

+++ AMCOS.PostgreSQL/migrations/006j_warehouse_populate_procs.sql
@@ -2091,7 +2091,10 @@ BEGIN
     FROM gs_subgroup a
         INNER JOIN gradelevel b
             ON a.gs_ses_payplan = b.gs_ses_payplan
-           AND a.gs_ses_gradelevel = b.gs_ses_gradelevel
+           -- a.gs_ses_gradelevel is smallint (from data.costs via uniquecosts); b comes
+           -- from xwalk.ppxwalkgradelevel which casts gs_ses_gradelevel to varchar(2).
+           -- Cast back to smallint for a numeric equality match (grades are numeric here).
+           AND a.gs_ses_gradelevel = b.gs_ses_gradelevel::smallint
            AND a.topayplan = b.topayplan;
 
     --## Step 3, add in the location equivalent
